Airflow_Data_Pipeline_for_Sentiment_Analysis-main/requirements.txt
finnhub-python
pymongo
sqlalchemy==1.4.46
textblob
apache-airflow==2.8.4
pandas==2.1.4
psycopg2-binary
python-dotenv